webpack build add babel-loader cache, optimize common chunk - remove webpack.common.js - fix index.pug
