+++ package.json
@@ -40,8 +40,8 @@
     "dev:client": "NODE_ENV=development webpack --config webpack.dev.js --progress",
     "dev:serve": "DEBUG=koa-*,-koa-views BABEL_ENV=server COMPILE_CLIENT=true NODE_ENV=development node server/server.js",
     "server": "DEBUG=koa-router BABEL_ENV=server NODE_ENV=development node server/server.js",
-    "dev:watch": "npm run dev:vendor && npm run dev:client -- --watch",
-    "dev": "npm run dev:vendor && npm run dev:client && npm run server",
+    "dev:watch": "npm run dev:client -- --watch",
+    "dev": "rimraf build && npm run dev:client && npm run server",
     "prod:build": "NODE_ENV=production webpack -p --config webpack.prod.js --progress",
     "prod:serve": "DEBUG=koa-router node dist/server.js",
     "prod": "rimraf dist && npm run prod:build && npm run prod:serve",
